Refactoring
Change log:
1. Refactored a large portion of the code
2. Fixed some of the bugs
